fixing tests by moving fixture
